// src/app/resolvingoferrors/models/unsolicited-error.ts
export interface UnSolicitedErrors {
    
    Reference: string;
    View: string;
    TelNo: string;
    Source: string;
    ErrorCount: string;
    ResolutionType: string;
    RequestStart: string;
    RequestEnd: string;
    Diff: string;
    Reference1: string;
    LatestUserComments: string;
    LatestCommentDate: string;
    
}

export interface InformationTable1 {
    Month: string;
    Resolve: string;
    Count: string;
}

export interface InformationTable2 {
    Month: string;
    New: string;
    Investigation: string;
    Governance: string;
    Port: string;
    PComp: string;
    Resolve: string;
    Other: string;
    
}